+++ servicio-transporte/src/main/java/com/backend/tpi_backend/servicio_transporte/repositories/TramoTipoRepository.java
@@ -2,5 +2,5 @@ package com.backend.tpi_backend.servicio_transporte.repositories;
 
 import com.backend.tpi_backend.servicio_transporte.model.TramoTipo;
 
-public interface TramoTipoRepository extends BaseRepository<TramoTipo, Long> {
+public interface TramoTipoRepository extends BaseRepository<TramoTipo, Integer> {
 }
